+++ packages/core/eg/runOtherCode.ts
@@ -19,24 +19,33 @@ await dev('测试与 run other code 相关的逻辑', async it => {
   });
 
   await it('测试等待', async () => {
+    const waiting = waitingTips({
+      // interval: 100,
+      prefix: 1,
+      canCtrlCExit: true,
+    });
     const result = await runOtherCode({
       // code: `sleep 20 && echo "11"`,
       code: `sleep 3 && echo "h" && sleep 1 && echo "hello" && sleep 2 && echo "11"`,
       printLog: true,
       cwd: '/a',
-      waiting: {
-        prefix: 2,
-        // interval: 100,
-        info: '65465464646489798798797897979879879879789789798798798798798797979src/readInput/dataStore.ts -> src/readInput/stdRemoveListener.ts -> src/readInput/pressCallFn.ts -> src/readInput/dataStore.ts98',
-      },
+      waiting,
+      // waiting: {
+      //   prefix: 2,
+      //   // interval: 100,
+      //   canCtrlCExit: true,
+      //   canCtrlDExit: true,
+      //   info: '65465464646489798798797897979879879879789789798798798798798797979src/readInput/dataStore.ts -> src/readInput/stdRemoveListener.ts -> src/readInput/pressCallFn.ts -> src/readInput/dataStore.ts98',
+      // },
     });
-    console.log(result);
+    console.log(result.isSIGINT);
   });
 
   await it.skip('测试单独的 awaiting', async () => {
     const result = waitingTips({
       // interval: 100,
       prefix: 1,
+      canCtrlCExit: true,
     });
     result.log('开始执行:', new Date().toLocaleString());
     await sleep(1000);
@@ -52,6 +61,10 @@ await dev('测试与 run other code 相关的逻辑', async it => {
     result.log('方法已注销:', new Date().toLocaleString());
     await sleep(2000);
     result.log('销毁后等待 2 秒:', new Date().toLocaleString());
+    result.run('从新激活');
+    result.run('从新激活');
+    result.run('从新激活');
+    result.run('从新激活');
     result.run('从新激活');
     result.log('重新激活 2 秒:', new Date().toLocaleString());
     await sleep(2000);