+++ packages/core/index.ts
@@ -23,13 +23,13 @@ export type {
   RunOtherCodeResult,
 } from './src/run-other-code';
 
-export type {
-  RunOtherCodeWaiting,
-  waitingTipsResult,
-  waitingTipsParams,
-} from './src/waiting';
+export type { RunOtherCodeWaiting, waitingTipsParams } from './src/waiting';
 
-export { waitingTipsPrefixStore, waitingTips } from './src/waiting';
+export {
+  waitingTipsPrefixStore,
+  waitingTips,
+  WaitingTipsResult,
+} from './src/waiting';
 
 export { getNpmPkgInfo, testNpmPackageExist } from './src/npmPkg';